M9: tester zip --x64 - the 64-bit exes beside the 32-bit pair (PR 3 of 3)

+++ source/gfx/prim.cpp
@@ -42,7 +42,7 @@ void	PrimInit()
 void	PrimDisplay()
 {
 	while(DrawSync(1));
-	if ((int)CurrPrim>=(int)EndPrim)
+	if ((int)(size_t)CurrPrim>=(int)(size_t)EndPrim)
 	{
 		ASSERT(!"PRIM OVERFLOW");
 	}
